Połączenie z bazą, wyświetlanie i dodawanie requestów
J/w,
Jeszcze do zrobienia:
- przypisanie pracowników do eventu
- poprawienie wczytywania/wyświetlaniu equipmentu sal (w test mode
wyświetlają się prawidłowo, więc podejrzewam, że jest jakiś błąd przy
pobieraniu ich z bazy)
- kosmetyka typu komunikaty po dodaniu requestu itp.

+++ IO - wtf is wrong/src/filter/RoomFilter.java
@@ -49,14 +49,14 @@ public class RoomFilter implements Filter<RoomEquipment> {
 			}
 
 			if (var.containsKey("MAX_PRICE")) {
-				Integer value = (Integer) var.get("MAX_PRICE");
+				Float value = (Float) var.get("MAX_PRICE");
 				if (room.getPrice() > value) {
 					continue;
 				}
 			}
 
 			if (var.containsKey("MIN_PRICE")) {
-				Integer value = (Integer) var.get("MIN_PRICE");
+				Float value = (Float) var.get("MIN_PRICE");
 				if (room.getPrice() < value) {
 					continue;
 				}
